Move the full CLI to the new jobs.py.
